related work: drop part about Vue, React & FRP
Time-varying signals is just another way of saying "state".  I think
the more significant difference is between incremental and discrete,
but even that's pretty hard to tease out.  For example, you could
argue React is incremental in the way it does diffing and updates.
You could also argue that observables are incremental if you use lots
of small observables.  Best not to dive into this IMO.